+++ allennlp/modules/vision/grid_embedder.py
@@ -68,7 +68,7 @@ class ResnetBackbone(GridEmbedder):
         super().__init__()
         from allennlp.common import detectron
 
-        flat_parameters = detectron.DetectronFlatParameters(
+        self.flat_parameters = detectron.DetectronFlatParameters(
             max_attr_per_ins=max_attr_per_ins,
             device=device,
             weights=weights,
@@ -79,10 +79,25 @@ class ResnetBackbone(GridEmbedder):
             width_per_group=width_per_group,
             depth=depth,
         )
+        self._pipeline_object = None
 
-        pipeline = detectron.get_pipeline_from_flat_parameters(flat_parameters, make_copy=False)
-        self.preprocessor = pipeline.model.preprocess_image
-        self.backbone = pipeline.model.backbone
+    @property
+    def _pipeline(self):
+        if self._pipeline_object is None:
+            from allennlp.common import detectron
+
+            self._pipeline_object = detectron.get_pipeline_from_flat_parameters(
+                self.flat_parameters, make_copy=False
+            )
+        return self._pipeline_object
+
+    @property
+    def preprocessor(self):
+        return self._pipeline.model.preprocess_image
+
+    @property
+    def backbone(self):
+        return self._pipeline.model.backbone
 
     def forward(self, images: FloatTensor, sizes: IntTensor) -> FloatTensor:
         images = [
@@ -99,3 +114,14 @@ class ResnetBackbone(GridEmbedder):
 
     def get_stride(self) -> int:
         return self.backbone.output_shape()["res4"].stride
+
+    def to(self, device):
+        if isinstance(device, int) or isinstance(device, torch.device):
+            if self._pipeline_object is not None:
+                self._pipeline_object.model.to(device)
+            if isinstance(device, torch.device):
+                device = device.index
+            self.flat_parameters = self.flat_parameters._replace(device=device)
+            return self
+        else:
+            return super().to(device)