+++ allennlp/predictors/__init__.py
@@ -9,3 +9,10 @@ a `Predictor` that wraps it.
 from allennlp.predictors.predictor import Predictor
 from allennlp.predictors.sentence_tagger import SentenceTaggerPredictor
 from allennlp.predictors.text_classifier import TextClassifierPredictor
+
+try:
+    from allennlp.predictors.vilbert_vqa import VilbertVqaPredictor
+    from allennlp.predictors.visual_entailment import VisualEntailmentPredictor
+except ImportError:
+    # vision-based predictors are not available if we don't have detectron.
+    pass